Bring settings to version control.
Note about .obsidian/community-plugins.json: In addition to "obsidian-shellcommands", the file also contains "hot-reaload". The latter plugin is only meant for plugin development purposes. Other users don't need to install it, and it's not even available in the community plugins list. It should not cause any problems to not have that plugin installed even though it's mentioned in the file.